+++ app/components/hooks/useTotalStakedTokens.ts
@@ -2,7 +2,7 @@ import { ethers } from "ethers";
 import { UseQueryResult, useQuery } from "@tanstack/react-query";
 import { IERC20__factory } from "@/typechain";
 import { SupportedNetwork, TokenRecord, supportedTokens } from "@/app/utils/types";
-import { getNetworkTokens, getNetworkProvider } from "@/app/constants";
+import { getNetworkTokens, getNetworkProvider } from "@/app/utils/constants";
 import { getTotalStakedBeacon } from "@/app/utils/api/totalStakedBeacon";
 
 async function getStrategyTvl(tokenAddress: `0x${string}`, strategyAddress: `0x${string}`, provider: ethers.Provider): Promise<number> {
